+++ release_notes.md
@@ -5,6 +5,7 @@ v0.6.0
 -----
 * Added address generation
 * Added sample to get a block by height
+* Added batching for blocks
 
 v0.5.0
 ------